requirement.txt
google-genai
fastapi
python-dotenv
pydantic
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-jose
pytest